backend/requirements.txt
sqlalchemy
fastapi
slowapi
uvicorn[standard]
pymongo
motor
itsdangerous
pymysql
requests
sshtunnel
stripe
python-dotenv
pyotp
pytest
pytest-mock
httpx
pytest-playwright
playwright
pytest-env
python-json-logger